+++ src/core/signals.rs
